Changed sorting functions so that they just use area from the  stack in the case of short lists and so avoid a malloc.  For  applications that use sorting heavily (e.g. set-based analysis), this  greately improves scalability for the MT engine, and should improve  speed a little for single threade applications as well.

+++ emu/init_xsb.c
@@ -1394,7 +1394,7 @@ void init_symbols(void)
   	if (status != 0) 
     		xsb_initialization_exit("Cannot determine thread stack size during system initialization");
 	else
-		printf( "Minimum thread stack size set to %d\n", stack_size ) ;
+	  printf( "Minimum thread stack size set to %d\n", (int) stack_size ) ;
   }
 #endif
   pthread_attr_setscope(&normal_attr_gl,PTHREAD_SCOPE_SYSTEM);
